requirements.txt
streamlit
requests
pandas
plotly
selenium
beautifulsoup4
python-dotenv
serpapi
google-generativeai
tabulate
